Move Mjolnir setup and configuration to a sensible place.
